+++ src/lib/index.ts
@@ -1,2 +1,3 @@
 export { PapaParseModule } from './papaparse.module';
 export { PapaParseService } from './papaparse.service';
+export * from './papaparse.interface';